騎手・調教師のIDをi32に統一

+++ ukeiba-common/src/scraper/jockey_profile.rs
@@ -8,13 +8,13 @@ use unicode_normalization::UnicodeNormalization;
 
 #[derive(Debug, Clone)]
 pub struct Page {
-    pub jockey_nar_id: i64,
+    pub jockey_nar_id: i32,
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 
 pub struct Data {
-    pub jockey_nar_id: i64,
+    pub jockey_nar_id: i32,
     pub name: String,
     pub kana: String,
     pub sex: String,
